Narrowing the visibility of some methods in ProtoUtil.
