+++ Core/src/test/java/org/tribuo/transform/transformations/MeanStdDevTest.java
@@ -158,8 +158,5 @@ public class MeanStdDevTest {
 
         Transformer tD = ProtoUtil.deserialize(tp);
         assertEquals(t, tD);
-        
-        assertEquals(TransformerProto.class, ProtoUtil.getSerializedClass(t));
-
-    }    
+    }
 }
