+++ Core/src/test/java/org/tribuo/transform/transformations/SimpleTransformTest.java
@@ -347,8 +347,6 @@ public class SimpleTransformTest {
 
         Transformer tD = ProtoUtil.deserialize(tp);
         assertEquals(t, tD);
-        
-        assertEquals(TransformerProto.class, ProtoUtil.getSerializedClass(t));
     }
 
 }
